requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.2
scikit-learn>=0.24.2
pillow>=8.2.0
opencv-python>=4.5.2
tqdm>=4.61.1
ipykernel>=6.0.1